requirements.txt
flask>=2.0.1
google-cloud-vision>=3.1.1
numpy>=1.24.0
flask-cors>=3.0.10
